+++ app/src/components/annotation/AnnotationLabel.tsx
@@ -1,14 +1,9 @@
 import { PropsWithChildren } from "react";
 import { css } from "@emotion/react";
 
-import { Flex, Icon, Icons, Text } from "@phoenix/components";
-import { assertUnreachable } from "@phoenix/typeUtils";
-import { formatFloat } from "@phoenix/utils/numberFormatUtils";
+import { AnnotationNameAndValue } from "@phoenix/components/annotation/AnnotationNameAndValue";
 
-import { AnnotationColorSwatch } from "./AnnotationColorSwatch";
-import { Annotation } from "./types";
-
-type AnnotationDisplayPreference = "label" | "score" | "none";
+import { Annotation, AnnotationDisplayPreference } from "./types";
 
 export const baseAnnotationLabelCSS = css`
   border-radius: var(--ac-global-dimension-size-50);
@@ -16,55 +11,19 @@ export const baseAnnotationLabelCSS = css`
   padding: var(--ac-global-dimension-size-50)
     var(--ac-global-dimension-size-100);
   transition: background-color 0.2s;
+  display: flex;
+  flex-direction: row;
+  gap: var(--ac-global-dimension-size-50);
   &[data-clickable="true"] {
     cursor: pointer;
     &:hover {
       background-color: var(--ac-global-color-grey-300);
     }
   }
-
   .ac-icon-wrap {
     font-size: 12px;
   }
 `;
-
-const textCSS = css`
-  display: flex;
-  align-items: center;
-  .ac-text {
-    display: inline-block;
-    max-width: 9rem;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-`;
-
-const getAnnotationDisplayValue = (
-  annotation: Annotation,
-  displayPreference: AnnotationDisplayPreference
-) => {
-  switch (displayPreference) {
-    case "label":
-      return (
-        annotation.label ||
-        (typeof annotation.score == "number" &&
-          formatFloat(annotation.score)) ||
-        "n/a"
-      );
-    case "score":
-      return (
-        (typeof annotation.score == "number" &&
-          formatFloat(annotation.score)) ||
-        annotation.label ||
-        "n/a"
-      );
-    case "none":
-      return "";
-    default:
-      assertUnreachable(displayPreference);
-  }
-};
 
 export function AnnotationLabel({
   annotation,
@@ -73,7 +32,6 @@ export function AnnotationLabel({
   className,
   children,
   clickable: _clickable,
-  showClickableIcon = true,
 }: PropsWithChildren<{
   annotation: Annotation;
   /**
@@ -82,11 +40,6 @@ export function AnnotationLabel({
    * clickable element (e.g. a dialog trigger, a link, etc).
    */
   clickable?: boolean;
-  /**
-   * When an annotation is clickable, this prop controls whether to show the click affordance icon.
-   * @default true
-   */
-  showClickableIcon?: boolean;
   onClick?: () => void;
   /**
    * The preferred value to display in the annotation label.
@@ -100,11 +53,6 @@ export function AnnotationLabel({
   className?: string;
 }>) {
   const clickable = _clickable ?? typeof onClick == "function";
-  const labelValue = getAnnotationDisplayValue(
-    annotation,
-    annotationDisplayPreference
-  );
-
   return (
     <div
       role={clickable ? "button" : undefined}
@@ -124,30 +72,11 @@ export function AnnotationLabel({
         }
       }}
     >
-      <Flex direction="row" gap="size-100" alignItems="center">
-        <AnnotationColorSwatch annotationName={annotation.name} />
-        <div css={textCSS}>
-          <Text weight="heavy" size="XS" color="inherit">
-            {annotation.name}
-          </Text>
-        </div>
-        {labelValue && (
-          <div
-            css={css(
-              textCSS,
-              css`
-                margin-left: var(--ac-global-dimension-100);
-              `
-            )}
-          >
-            <Text size="XS">{labelValue}</Text>
-          </div>
-        )}
-        {children}
-        {clickable && showClickableIcon ? (
-          <Icon svg={<Icons.ArrowIosForwardOutline />} />
-        ) : null}
-      </Flex>
+      <AnnotationNameAndValue
+        annotation={annotation}
+        displayPreference={annotationDisplayPreference}
+      />
+      {children}
     </div>
   );
 }
