+++ app/src/components/annotation/AnnotationTooltip.tsx
@@ -1,16 +1,8 @@
 import { ReactNode } from "react";
 
-import {
-  Flex,
-  RichTooltip,
-  Text,
-  TooltipTrigger,
-  TriggerWrap,
-  View,
-} from "@phoenix/components";
-import { Truncate } from "@phoenix/components/utility/Truncate";
-import { floatFormatter } from "@phoenix/utils/numberFormatUtils";
+import { RichTooltip, TooltipTrigger, TriggerWrap } from "@phoenix/components";
 
+import { AnnotationTooltipContent } from "./AnnotationTooltipContent";
 import { Annotation } from "./types";
 
 /**
@@ -19,97 +11,17 @@ import { Annotation } from "./types";
 export function AnnotationTooltip({
   annotation,
   children,
-  extra,
-  layout = "vertical",
-  leadingExtra,
 }: {
   leadingExtra?: ReactNode;
   annotation: Annotation;
   children: ReactNode;
-  layout?: "horizontal" | "vertical";
   extra?: ReactNode;
 }) {
   return (
     <TooltipTrigger delay={500}>
       <TriggerWrap>{children}</TriggerWrap>
       <RichTooltip offset={3}>
-        <Flex
-          direction={layout === "horizontal" ? "row" : "column"}
-          alignItems="start"
-        >
-          <Flex
-            direction="column"
-            gap="size-100"
-            height="100%"
-            justifyContent="space-between"
-          >
-            {leadingExtra}
-            <View>
-              <Text weight="heavy" color="inherit" size="L" elementType="h3">
-                {annotation.name}
-              </Text>
-              <View paddingTop="size-100" minWidth="150px">
-                <Flex direction="row" justifyContent="space-between">
-                  <Text weight="heavy" color="inherit">
-                    label
-                  </Text>
-                  <Text color="inherit" title={annotation.label ?? undefined}>
-                    <Truncate maxWidth="200px">
-                      {annotation.label || "--"}
-                    </Truncate>
-                  </Text>
-                </Flex>
-                <Flex direction="row" justifyContent="space-between">
-                  <Text weight="heavy" color="inherit">
-                    score
-                  </Text>
-                  <Text color="inherit">
-                    {floatFormatter(annotation.score)}
-                  </Text>
-                </Flex>
-                {annotation.annotatorKind ? (
-                  <Flex direction="row" justifyContent="space-between">
-                    <Text weight="heavy" color="inherit">
-                      annotator kind
-                    </Text>
-                    <Text color="inherit">{annotation.annotatorKind}</Text>
-                  </Flex>
-                ) : null}
-              </View>
-              {annotation.explanation ? (
-                <View paddingTop="size-50">
-                  <Flex direction="column">
-                    <Text weight="heavy" color="inherit">
-                      explanation
-                    </Text>
-                    <View maxHeight="300px" overflow="auto">
-                      <Text color="inherit">{annotation.explanation}</Text>
-                    </View>
-                  </Flex>
-                </View>
-              ) : null}
-              {annotation.createdAt ? (
-                <Flex
-                  direction="row"
-                  justifyContent="space-between"
-                  wrap="wrap"
-                  gap="size-100"
-                >
-                  <Text weight="heavy" color="inherit">
-                    created at
-                  </Text>
-                  <Text color="inherit">
-                    {new Date(annotation.createdAt)
-                      .toLocaleString()
-                      .split(",")
-                      .join(",\n")}
-                  </Text>
-                </Flex>
-              ) : null}
-            </View>
-          </Flex>
-          {extra}
-        </Flex>
+        <AnnotationTooltipContent annotation={annotation} />
       </RichTooltip>
     </TooltipTrigger>
   );
